simploxide-client: Make event stream a named type
Working with the `impl Stream` was unergonomic not only because it was a
nightmare to type out but also because it was quite easy to forget to
propagate the Unpin bound.

Besides turning the event stream into a non-existential type this patch
adds some public typedef exports to `simploxide-core` to simplify
wrapping the raw event queue into higher-level types.